[2단계 - 주문 기능 구현] 채채(신채원) 미션 제출합니다. (#95)
* feat: cors 설정

* feat: exposedHeaders Location 추가

* refactor: exposedHeaders 허용 범위 수정

* feat: db 테이블 추가 및 데이터 추가

* feat: 장바구니 저장 기능 구현

* feat: 장바구니 회원의 모든 주문 조회 기능 구현

* feat: 주문 저장 기능 구현

* fix: 저장기능 오류 수정

* refactor: 주문저장기능 구현 시 주문아이템 저장 기능 추가

* feat: 회원 포인트 조회 기능 추가

* refactor: order 도메인에 point 추가

* fix: order 저장 api 오류 수정

* style:변수명 수정

* refactor: 회원에게 추가되는 포인트 로직 수정

* fix: 주문 목록들 가져오는 로직 null값 오류 수정

* refactor: 응답dto에 데이터 수정

* refactor: 포인트 퍼센트 전략 인터페이스 분리

* feat: 주문 정책 가져오기 구현

* refactor 파일 이동 및 파일 이름 수정

* refactor: 등록 api dto 생성하여 반환값 수정

* feat: 예외처리 구현 및 핸들러 구현

* feat: 데이터베이스 서버 연동 구현

* feat: 데이터베이스 서버 연동 구현

* fix: 빌드파일 수정

* fix: 빌드파일 수정

* refactor: 사용하지 않는 상수 수정

* fix: sql문 오류 수정

* style: 오류 메서드 네이밍 수정

* refactor: 주문 예외처리 도메인으로 분리

* refactor: shippingFee 도메인 수정

* refactor: 공동된 예외처리 묶어서 리팩토링

* test: 테스트 추가

+++ src/main/java/cart/domain/Product.java
@@ -3,16 +3,16 @@ package cart.domain;
 public class Product {
     private Long id;
     private String name;
-    private int price;
+    private Long price;
     private String imageUrl;
 
-    public Product(String name, int price, String imageUrl) {
+    public Product(String name, Long price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
     }
 
-    public Product(Long id, String name, int price, String imageUrl) {
+    public Product(Long id, String name, Long price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -27,7 +27,7 @@ public class Product {
         return name;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
